+++ src/asn.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: asn.cc,v 1.115 2007/09/21 11:41:52 amosjeffries Exp $
+ * $Id: asn.cc,v 1.115.2.1 2008/02/25 23:08:50 amosjeffries Exp $
  *
  * DEBUG: section 53    AS Number handling
  * AUTHOR: Duane Wessels, Kostas Anagnostakis
@@ -89,7 +89,7 @@ struct ASState
     store_client *sc;
     HttpRequest *request;
     int as_number;
-    off_t offset;
+    int64_t offset;
     int reqofs;
     char reqbuf[AS_REQBUF_SZ];
     bool dataRead;
@@ -299,7 +299,7 @@ asHandleReply(void *data, StoreIOBuffer result)
      */
     s = buf;
 
-    while (s - buf < (off_t)(result.length + asState->reqofs) && *s != '\0') {
+    while ((size_t)(s - buf) < result.length + asState->reqofs && *s != '\0') {
         while (*s && xisspace(*s))
             s++;
